+++ package.json
@@ -4,8 +4,10 @@
   "description": "Climate_Cooperatives Open Source API",
   "main": "src/app.js",
   "scripts": {
-    "start": "node src/local.js",
-    "test": "jest"
+    "test": "jest",
+    "build": "esbuild ./src/local.ts --bundle --platform=node --outdir=dist",
+    "start:local": "node dist/local.js",
+    "build:start:local": "npm run build & npm run start:local"
   },
   "author": "Climate_Cooperative",
   "license": "ISC",
@@ -23,6 +25,7 @@
     "@babel/core": "^7.24.4",
     "@babel/preset-env": "^7.24.4",
     "@babel/preset-typescript": "^7.24.1",
+    "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
     "@types/jest": "^29.5.12",
     "@types/node": "^20.12.12",
